refactor: move profiles into library
