+++ src/lib.rs
@@ -1,7 +1,9 @@
 mod generation;
 mod generation_set;
 mod job;
+mod profiles;
 
 pub use generation::Generation;
 pub use generation_set::GenerationSet;
 pub use job::Job;
+pub use profiles::Profile;
